refactor(httpauth) move validator() to api_proxy
